nextCode2 updated to be more generic
I've updated the nextCode2 test code to generate length from 1 to whatever number we want.

As all goes first now...

The cut off is always 3 letters from the end.
So if the current test length is 18 characters long, the sequential portion will be 15 and the last 3 will be recursive.
This allows us to break up the work to threads.